[#84] Rename Vocabulary.js to vocabulary.js to follow the naming convention for javascript files
