+++ src/components/ScriptFunctionSelection.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Dropdown } from "semantic-ui-react";
-import { FUNCTION_URI, FUNCTION_NAME } from "../constants/Vocabulary.js";
+import { FUNCTION_URI, FUNCTION_NAME } from "../constants/vocabulary.js";
 import Rest from "../rest/Rest.jsx";
 
 class ScriptFunctionSelection extends React.Component {
