+++ src/constants/vocabulary.js
